refactor: separate mememaker component and styles
