+++ WT4Q/src/app/tools/mememaker/css/mememaker.module.css
@@ -1,0 +1,3 @@
+.rimLight {
+  filter: drop-shadow(0 0 6px rgba(255, 255, 255, 0.25));
+}
